requirements.txt
streamlit
numpy
matplotlib
opencv-python-headless==4.8.1.78
scikit-learn
seaborn
tensorflow
Pillow